+++ tisbi-rip-server/database/migrations/2025_05_22_183124_create_bonuses_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->foreignId('employee_id');
             $table->date('date');
-            $table->float('sum', 2);
+            $table->decimal('sum');
         });
     }
